Change harga and nilai fields to decimal for PO items
Updated PurchaseOrderItem model and related logic to use decimal type for 'harga' and 'nilai' fields to prevent overflow with large numbers. Added migrations to alter the database schema accordingly. Adjusted casting and calculation logic in controllers and models. Also improved UI feedback and field requirements in Mitra and Admin Vue pages.

+++ app/Models/PurchaseOrder.php
@@ -71,7 +71,8 @@ class PurchaseOrder extends Model
         if (!$this->relationLoaded('items')) {
             $this->load('items');
         }
-        return (int) ($this->items->sum('nilai') ?? 0);
+        // Jangan cast ke int untuk menghindari overflow pada angka besar
+        return $this->items->sum('nilai') ?? 0;
     }
 
     // Accessor untuk mendapatkan jenis komoditas yang tepat
